pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
tweepy = "^4.10.0"
requests = "^2.28.1"
Pillow = "^9.2.0"
imageio = "^2.19.5"
matplotlib = "^3.5.2"
Flask = "^2.1.3"
schedule = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"